Code improvements

+++ src/main/java/tech/jhipster/lite/shared/error/domain/Assert.java
@@ -458,7 +458,7 @@ public final class Assert {
     public IntegerAsserter min(int minValue) {
       notNull(field, value);
 
-      if (value.intValue() < minValue) {
+      if (value < minValue) {
         throw NumberValueTooLowException.builder().field(field).minValue(String.valueOf(minValue)).value(String.valueOf(value)).build();
       }
 
@@ -479,7 +479,7 @@ public final class Assert {
     public IntegerAsserter max(int maxValue) {
       notNull(field, value);
 
-      if (value.intValue() > maxValue) {
+      if (value > maxValue) {
         throw NumberValueTooHighException.builder().field(field).maxValue(String.valueOf(maxValue)).value(String.valueOf(value)).build();
       }
 
@@ -527,7 +527,7 @@ public final class Assert {
     public LongAsserter min(long minValue) {
       notNull(field, value);
 
-      if (value.longValue() < minValue) {
+      if (value < minValue) {
         throw NumberValueTooLowException.builder().field(field).minValue(String.valueOf(minValue)).value(String.valueOf(value)).build();
       }
 
@@ -548,7 +548,7 @@ public final class Assert {
     public LongAsserter max(long maxValue) {
       notNull(field, value);
 
-      if (value.longValue() > maxValue) {
+      if (value > maxValue) {
         throw NumberValueTooHighException.builder().field(field).maxValue(String.valueOf(maxValue)).value(String.valueOf(value)).build();
       }
 
@@ -609,7 +609,7 @@ public final class Assert {
     public FloatAsserter min(float minValue) {
       notNull(field, value);
 
-      if (value.floatValue() < minValue) {
+      if (value < minValue) {
         throw tooLow(minValue);
       }
 
@@ -630,7 +630,7 @@ public final class Assert {
     public FloatAsserter over(float floor) {
       notNull(field, value);
 
-      if (value.floatValue() <= floor) {
+      if (value <= floor) {
         throw tooLow(floor);
       }
 
@@ -655,7 +655,7 @@ public final class Assert {
     public FloatAsserter max(float maxValue) {
       notNull(field, value);
 
-      if (value.floatValue() > maxValue) {
+      if (value > maxValue) {
         throw tooHigh(maxValue);
       }
 
@@ -676,7 +676,7 @@ public final class Assert {
     public FloatAsserter under(float ceil) {
       notNull(field, value);
 
-      if (value.floatValue() >= ceil) {
+      if (value >= ceil) {
         throw tooHigh(ceil);
       }
 
@@ -741,7 +741,7 @@ public final class Assert {
     public DoubleAsserter min(double minValue) {
       notNull(field, value);
 
-      if (value.doubleValue() < minValue) {
+      if (value < minValue) {
         throw tooLow(minValue);
       }
 
@@ -762,7 +762,7 @@ public final class Assert {
     public DoubleAsserter over(double floor) {
       notNull(field, value);
 
-      if (value.doubleValue() <= floor) {
+      if (value <= floor) {
         throw tooLow(floor);
       }
 
@@ -787,7 +787,7 @@ public final class Assert {
     public DoubleAsserter max(double maxValue) {
       notNull(field, value);
 
-      if (value.doubleValue() > maxValue) {
+      if (value > maxValue) {
         throw tooHigh(maxValue);
       }
 
@@ -808,7 +808,7 @@ public final class Assert {
     public DoubleAsserter under(double ceil) {
       notNull(field, value);
 
-      if (value.doubleValue() >= ceil) {
+      if (value >= ceil) {
         throw tooHigh(ceil);
       }
